Fixing compilation issues with fps counter and ecs orchestrator.

+++ include/re/utils/fps_counter.cpp
@@ -1,6 +1,6 @@
 #include "fps_counter.h"
 
-FPSCounter::FPSCounter() : timer(new Timer(1000, true)) {
+FPSCounter::FPSCounter() : timer(new Timer((Uint32) 1000, true)) {
     timer->Start();
 }
